(PC-16819) refactor(offer): extract ios and android tests for share button from OfferHeader
